refactorisation of api_service and fix refreshToken handle
- Implemented AuthService for user login, registration, logout, and account deletion.
- Created FeedbackService to handle user feedback submissions with screenshots.
- Developed FriendService for managing friend requests and friendships.
- Added GinkoService to fetch schedules based on location.
- Introduced RestaurantService for retrieving menus, sectors, and user interactions in restaurants.
- Established SocketService for managing chat messages in chat rooms.
- Built UserService for user data retrieval and updates including password, status, username, and profile picture.
